WIP: Optimize differential testing

+++ packages/contracts-bedrock/contracts/test/Encoding.t.sol
@@ -64,7 +64,7 @@ contract Encoding_Test is CommonTest {
         uint64 _gas,
         bool isCreate,
         bytes memory _data,
-        uint256 _logIndex
+        uint64 _logIndex
     ) external {
         Types.UserDepositTransaction memory t = Types.UserDepositTransaction(
             _from,
